PRSD-1246: Add EICR update CYA page
