+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/PropertyComplianceCheckAnswersPage.kt
@@ -1,23 +1,14 @@
 package uk.gov.communities.prsdb.webapp.forms.pages
 
-import kotlinx.datetime.DatePeriod
-import kotlinx.datetime.plus
 import org.springframework.web.servlet.ModelAndView
-import uk.gov.communities.prsdb.webapp.constants.EICR_VALIDITY_YEARS
 import uk.gov.communities.prsdb.webapp.constants.EPC_ACCEPTABLE_RATING_RANGE
-import uk.gov.communities.prsdb.webapp.constants.enums.EicrExemptionReason
 import uk.gov.communities.prsdb.webapp.forms.JourneyData
+import uk.gov.communities.prsdb.webapp.forms.pages.cya.EicrSummaryRowsFactory
 import uk.gov.communities.prsdb.webapp.forms.pages.cya.GasSafetySummaryRowsFactory
 import uk.gov.communities.prsdb.webapp.forms.steps.PropertyComplianceStepId
 import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getAcceptedEpcDetails
 import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getDidTenancyStartBeforeEpcExpiry
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getEicrExemptionOtherReason
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getEicrExemptionReason
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getEicrIssueDate
 import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getEpcExemptionReason
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasCompletedEicrExemptionConfirmation
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasCompletedEicrExemptionMissing
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasCompletedEicrUploadConfirmation
 import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasCompletedEpcExemptionConfirmation
 import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasCompletedEpcExpired
 import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasCompletedEpcMissing
@@ -45,6 +36,12 @@ class PropertyComplianceCheckAnswersPage(
             gasSafetyStartingStep = PropertyComplianceStepId.GasSafety,
             changeExemptionStep = PropertyComplianceStepId.GasSafetyExemption,
         )
+    val eicrDataFactory =
+        EicrSummaryRowsFactory(
+            doesDataHaveEicr = { data -> data.getHasEICR()!! },
+            eicrSafetyStartingStep = PropertyComplianceStepId.EICR,
+            changeExemptionStep = PropertyComplianceStepId.EicrExemption,
+        )
 
     override fun addPageContentToModel(
         modelAndView: ModelAndView,
@@ -52,21 +49,10 @@ class PropertyComplianceCheckAnswersPage(
     ) {
         modelAndView.addObject("propertyAddress", propertyAddressProvider())
         modelAndView.addObject("gasSafetyData", gasSafetyDataFactory.createRows(filteredJourneyData))
-        modelAndView.addObject("eicrData", getEicrData(filteredJourneyData))
+        modelAndView.addObject("eicrData", eicrDataFactory.createRows(filteredJourneyData))
         modelAndView.addObject("epcData", getEpcData(filteredJourneyData))
         modelAndView.addObject("responsibilityData", getResponsibilityData(filteredJourneyData))
     }
-
-    private fun getEicrData(filteredJourneyData: JourneyData) =
-        mutableListOf<SummaryListRowViewModel>()
-            .apply {
-                add(getEicrStatusRow(filteredJourneyData))
-                if (filteredJourneyData.getHasEICR()!!) {
-                    addAll(getEicrDetailRows(filteredJourneyData))
-                } else {
-                    add(getEicrExemptionRow(filteredJourneyData))
-                }
-            }.toList()
 
     private fun getEpcData(filteredJourneyData: JourneyData) =
         mutableListOf<SummaryListRowViewModel>()
@@ -99,57 +85,6 @@ class PropertyComplianceCheckAnswersPage(
                 actionValue = "forms.links.view",
             ),
         )
-
-    private fun getEicrStatusRow(filteredJourneyData: JourneyData): SummaryListRowViewModel {
-        val fieldValue =
-            // TODO PRSD-976: Add link to gas safety cert (or appropriate message if virus scan failed)
-            if (filteredJourneyData.getHasCompletedEicrUploadConfirmation()) {
-                "forms.checkComplianceAnswers.eicr.download"
-            } else if (filteredJourneyData.getHasCompletedEicrExemptionConfirmation()) {
-                "forms.checkComplianceAnswers.certificate.notRequired"
-            } else if (filteredJourneyData.getHasCompletedEicrExemptionMissing()) {
-                "forms.checkComplianceAnswers.certificate.notAdded"
-            } else {
-                "forms.checkComplianceAnswers.certificate.expired"
-            }
-
-        return SummaryListRowViewModel.forCheckYourAnswersPage(
-            "forms.checkComplianceAnswers.eicr.certificate",
-            fieldValue,
-            PropertyComplianceStepId.EICR.urlPathSegment,
-        )
-    }
-
-    private fun getEicrDetailRows(filteredJourneyData: JourneyData): List<SummaryListRowViewModel> {
-        val issueDate = filteredJourneyData.getEicrIssueDate()!!
-        return listOf(
-            SummaryListRowViewModel.forCheckYourAnswersPage(
-                "forms.checkComplianceAnswers.certificate.issueDate",
-                issueDate,
-                PropertyComplianceStepId.EicrIssueDate.urlPathSegment,
-            ),
-            SummaryListRowViewModel.forCheckYourAnswersPage(
-                "forms.checkComplianceAnswers.certificate.validUntil",
-                issueDate.plus(DatePeriod(years = EICR_VALIDITY_YEARS)),
-                null,
-            ),
-        )
-    }
-
-    private fun getEicrExemptionRow(filteredJourneyData: JourneyData): SummaryListRowViewModel {
-        val fieldValue: Any =
-            when (val exemptionReason = filteredJourneyData.getEicrExemptionReason()) {
-                null -> "commonText.none"
-                EicrExemptionReason.OTHER -> listOf(exemptionReason, filteredJourneyData.getEicrExemptionOtherReason())
-                else -> exemptionReason
-            }
-
-        return SummaryListRowViewModel.forCheckYourAnswersPage(
-            "forms.checkComplianceAnswers.certificate.exemption",
-            fieldValue,
-            PropertyComplianceStepId.EicrExemption.urlPathSegment,
-        )
-    }
 
     private fun getEpcStatusRow(filteredJourneyData: JourneyData): SummaryListRowViewModel {
         val fieldValue =
